scheduler/requirements.txt
PyYAML==6.0.1
pika==1.3.2
schedule==1.2.2